refactor: use embedded approach to reduce complexity by removing unnecessary IDs
Removed IDs from tables that do not require frequent joins. Adopted an embedded approach to simplify the database schema and reduce complexity.
