--- src/main/java/app/cardcapture/template/repository/PhraseRepository.java
@@ -1,9 +1,0 @@
-package app.cardcapture.template.repository;
-
-import app.cardcapture.template.domain.Phrase;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PhraseRepository extends JpaRepository<Phrase, Long> {
-}
